requirements.txt
python-dotenv
wordcloud
gensim
pyLDAvis
lda
tweetopic[viz]
tweetopic
fasttext
contractions
textblob_de
bertopic
pandas_profiling
openpyxl
python-dotenv
streamlit
pyodbc
datetime
plotly.express
nbformat
kaleido
